// extensions/bangumi/src/types/generated.ts
export interface paths {
  "/calendar": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /** 每日放送 */
    get: operations["getCalendar"]
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
  "/v0/search/subjects": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    get?: never
    put?: never
    /**
     * 条目搜索
     * @description ## 实验性 API， 本 schema 和实际的 API 行为都可能随时发生改动
     *
     *     目前支持的筛选条件包括:
     *     - `type`: 条目类型，参照 `SubjectType` enum， `或`。
     *     - `tag`: 标签，可以多次出现。`且` 关系。
     *     - `air_date`: 播出日期/发售日期。`且` 关系。
     *     - `rating`: 用于搜索指定评分的条目。`且` 关系。
     *     - `rating_count`: 用于按照评分人数筛选条目。`且` 关系。
     *     - `rank`: 用于搜索指定排名的条目。`且` 关系。
     *     - `nsfw`: 使用 `include` 包含NSFW搜索结果。默认排除搜索NSFW条目。无权限情况下忽略此选项，不会返回NSFW条目。
     *
     *     不同筛选条件之间为 `且`
     */
    post: operations["searchSubjects"]
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
  "/v0/search/characters": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    get?: never
    put?: never
    /**
     * 角色搜索
     * @description ## 实验性 API， 本 schema 和实际的 API 行为都可能随时发生改动
     *
     *     目前支持的筛选条件包括:
     *     - `nsfw`: 使用 `include` 包含NSFW搜索结果。默认排除搜索NSFW条目。无权限情况下忽略此选项，不会返回NSFW条目。
     */
    post: operations["searchCharacters"]
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
// ...
  "/v0/subjects/{subject_id}": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /**
     * 获取条目
     * @description cache with 300s
     */
    get: operations["getSubjectById"]
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
// ...
  "/v0/subjects/{subject_id}/characters": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /** Get Subject Characters */
    get: operations["getRelatedCharactersBySubjectId"]
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
  "/v0/subjects/{subject_id}/subjects": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /** Get Subject Relations */
    get: operations["getRelatedSubjectsBySubjectId"]
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
// ...
  "/v0/characters/{character_id}": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /**
     * Get Character Detail
     * @description cache with 60s
     */
    get: operations["getCharacterById"]
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
  "/v0/characters/{character_id}/image": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /** Get Character Image */
    get: operations["getCharacterImageById"]
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
  "/v0/characters/{character_id}/subjects": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /** get character related subjects */
    get: operations["getRelatedSubjectsByCharacterId"]
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
// ...
  "/v0/me": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /**
     * Get User
     * @description 返回当前 Access Token 对应的用户信息
     */
    get: operations["getMyself"]
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
  "/v0/users/{username}/collections": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /**
     * 获取用户收藏
     * @description 获取对应用户的收藏，查看私有收藏需要access token。
     */
    get: operations["getUserCollectionsByUsername"]
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
  "/v0/users/{username}/collections/{subject_id}": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /**
     * 获取用户单个条目收藏
     * @description 获取对应用户的收藏，查看私有收藏需要 access token
     */
    get: operations["getUserCollection"]
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
  "/v0/users/-/collections/{subject_id}": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    get?: never
    put?: never
    /**
     * 新增或修改用户单个条目收藏
     * @description 修改条目收藏状态, 如果不存在则创建，如果存在则修改
     *
     *     由于直接修改剧集条目的完成度可能会引起意料之外效果，只能用于修改书籍类条目的完成度。
     *
     *     方法的所有请求体字段均可选
     */
    post: operations["postUserCollection"]
    delete?: never
    options?: never
    head?: never
    /**
     * 修改用户单个收藏
     * @description 修改条目收藏状态
     *
     *     由于直接修改剧集条目的完成度可能会引起意料之外效果，只能用于修改书籍类条目的完成度。
     *
     *     PATCH 方法的所有请求体字段均可选
     */
    patch: operations["patchUserCollection"]
    trace?: never
  }
  "/v0/users/-/collections/{subject_id}/episodes": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /** 章节收藏信息 */
    get: operations["getUserSubjectEpisodeCollection"]
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    /**
     * 章节收藏信息
     * @description 同时会重新计算条目的完成度
     */
    patch: operations["patchUserSubjectEpisodeCollection"]
    trace?: never
  }
  "/v0/users/-/collections/-/episodes/{episode_id}": {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /** 章节收藏信息 */
    get: operations["getUserEpisodeCollection"]
    /** 更新章节收藏信息 */
    put: operations["putUserEpisodeCollection"]
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
// ...
}